Migrate package to null safety
This needed some small modifications because the List<E>(int)
constructor is no longer available. Instead, we create an empty list and
grow it up to capacity, at which point we switch over to overwriting
elements.

I also had to remove the dependency on the coverage package, because
it is not null safe yet (but looks like it will be shortly, based
on this commit https://github.com/dart-lang/coverage/commit/85141b7ba4dc3d715865e689926bef034aa85548).